requirements.txt
altgraph==0.17.4
crc==7.0.0
darkdetect==0.7.1
et-xmlfile==1.1.0
numpy==2.1.0
packaging==24.1
pefile==2023.2.7
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyqtdarktheme==2.1.0
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pywin32-ctypes==0.2.3
shiboken6==6.7.2
